Updated logging messages, refactored code, updated tests
